fix: relocation theme file
